Move src/pages/enspect/datails components
Move src/pages/enspect/datails components
to src/components/details
