+++ src/components/details/data-component.tsx
