+++ src/components/details/link-component.tsx
