+++ src/components/details/profile-card.tsx
