+++ src/pages/enspect/details/[id].tsx
@@ -1,9 +1,9 @@
 import PageLayout from "../../../components/layout/page";
 import { SimpleGrid, chakra } from "@chakra-ui/react";
-import FallBack from "./fallback";
-import ProfileCard from "./profile-card";
+import FallBack from "../../../components/details/fallback";
+import ProfileCard from "../../../components/details/profile-card";
 import SearchBar from "../../../components/input/enspect-search-bar";
-import EnsSearchError from "./ens-search-error";
+import EnsSearchError from "../../../components/details/ens-search-error";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { GetStaticProps, GetStaticPaths } from "next";
 import { useRouter } from "next/router";
